+++ UserEcho for IOS test app/UserEcho/noInternetConnectionVC.h
@@ -1,0 +1,18 @@
+//
+//  noInternetConnectionVC.h
+//  UserEcho for IOS test app
+//
+//  Created by Sergey Stukov on 8/26/13.
+//  Copyright (c) 2013 UserEcho. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface noInternetConnectionVC : UIViewController
+{
+    IBOutlet UINavigationBar* navBar;
+}
+
+-(IBAction)exit;
+
+@end